fix: trustworthy assessment submission + audit P2-P4 (extended T3)
Submission/grading trust (T3, audit P1):
- startAssessment/submitAssessment are now async and persisted-first:
  the success toast fires only after the backend saved; a failed submit PUT
  falls back to creating the full row; failures surface with retry paths
  instead of silently losing answers. TakeQuiz shows a dedicated
  'Couldn't start the attempt / Retry' view when the in_progress record
  can't be persisted.
- Unmount auto-submit no longer fires an async fetch during page teardown
  (browsers cancel it); it saves a DraftService snapshot instead, restored
  on resume. Auto-submit is double-fire-guarded and draft-falls back.
- TakeCoding + builder draft/publish: same await/catch treatment.
- Auto-graded percentage uses Math.floor to match the backend integer
  column (no more 67->66 flip after refresh) (P3).
- api.js: removed dead AuthService ('/iam' endpoints never existed) (S5);
  added DraftService.deleteDraft.

Answer-key privacy (P2):
- New backend GET /api/v1/assessments/student returns assessments with
  correctAnswer/explanation stripped; LMSContext loads it for student
  sessions. Assessments cache is keyed per viewer role so a stripped
  student copy can never poison a trainer/admin session.
- DraftController: DELETE endpoint for clearing drafts.

Float-safe marks (P3, storage): Answer.marksAwarded + Submission.score
Integer -> Double (psql: double precision), aligned StudentReportRowDto, so
negative-marking decimals (-2.5) persist exactly as graded.

Real progress on courses page (P4): cards now derive completion from
GET /progress/course/{id}; this also makes 'View Certificate' reachable
for genuinely completed courses (previously always 0% => dead button).
Requires X-User-Id fix: fetchApi now sends the logged-in user's id from
the session instead of hardcoded TEMPORARY_STUDENT_ID (largely satisfies
backlog X2).

Verified live through the gateway: student view has no correctAnswer while
trainer view keeps it; -2.5/17.5 round-trip through PostgreSQL; draft
save/get/delete cycle; full submit flow with server question ids.
Frontend build + assessment-service mvn install green.

+++ backend/assessment-service/src/main/java/com/xebia/assessmentservice/model/Answer.java
@@ -16,7 +16,7 @@ public class Answer {
     private String questionId;
     @Column(columnDefinition = "TEXT")
     private String answer;
-    private Integer marksAwarded;
+    private Double marksAwarded;
     @Column(columnDefinition = "TEXT")
     private String remarks;
 }
